+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - [0.1.0 - 2024-03-02](#010---2024-03-02)
 - [0.1.1 - 2024-03-02](#011---2024-03-02)
+- [1.0.0 - 2024-03-02](#100---2024-03-02)
 
 ## [0.1.0](https://github.com/jtompkin/pyclacker/releases/tag/v0.1.0) - 2024-03-02
 
@@ -18,8 +19,23 @@ Minor patch for program input
 
 ### Added
 
-- square root word (sqrt)
+- square root word: `sqrt`
 
 ### Fixed
 
-- Providing program on command line now did not always work
+- Providing program on command line did not always work
+
+## [1.0.0](https://github.com/jtompkin/pyclacker/releases/tag/v1.0.0) - 2024-03-02
+
+Big boy release
+
+### Added
+
+- Support for providing custom words in file
+- Word definition using '=' in interactive mode
+- pi word: `pi`
+
+### Changed
+
+- Now prints stack after last push in command
+- Entering '\\' no longer exits interactive mode (Use \<Ctrl-d>)